WIP: Set up message chain and object types
I'm setting up specific modules that a generic erlmud_object module
will use to interact with the data belonging to the object.

e.g. We can use erlmud_player to create a State object that will hold
everything pertaining to a player, and then we can pass all operations
to that module with that player State. The generic object gen_server
will handle all the message chain operations.