WIP: create the world
erlmud_world takes a list of properties and creates a list of
gen_servers. It tracks each gen_server's pid with the ID of the original
proplist. Once all the gen_servers are created it uses the map of ids to
pids to fill in all the properties of the gen_servers with the pids of
that match their IDs.

This allows us to forward reference gen_servers as members of fields
before we've instantiated them.

e.g.

{erlmud_room, room1, [{players, [player1]}]},
{erlmud_player, player1, [{room, room1}]}.

room1 can reference player1 even though player1 hasn't been instantiated
as a gen_server yet.

Any reference to a property that doesn't line up to a gen_server will be
ignored.