WIP: fixing premature message chain end
I was checking for an empty message chain before grabbing the processes
connected to the current object, so the chain was always empty.

Once I figured this out I needed to track the old messages (i.e. that
have already been handled) so that I didn't get a loops in the graph.

Now I'm using ordsets to manage old procs (processes that have handled
this message) and new procs (that still need to handle the message).

I'm also working linking up rooms with exits so that erlmud_object can
populate an exit with room processes instead of names and so erlmud_exit
can generate a list of processes.

erlmud_exit will store each room process in a 'rooms' property as a
tuple with the exit name: e.g. [{rooms, [{s, room1}, {n, room2}]}].