Move player from room to room
gen_server:cast(Room1, {attempt, {move, Player1, Room1, Room2}, Procs}).

Or,

P = erlmud_world:init(). erlmud_world:move(P).

... in the shell.

Procs in the gen_server:cast are the "Previous" processes, that have
already processed (and passed) the event, the "Next" processes, that
need to process the event and the "Subscribed" processes that want to be
notified about the result (pass or fail).