WIP: find exit -> move
Start a move by saying {move, Obj, Source, Exit} where exit is an atom
and then if the Source has an exit with a room through exit Exit then
translate the message to be {move, Obj, Source, Dest}.

If the message chain ends before an exit attaching Source to Dest
through exit direction Exit (which must be difference from the exit
through which Source is reached) then the last object in the chain can
fail the message. In this case the originating process (which has
subscribed to the result) can return "No exit 'blah'".

If the message does get translated then the originating process should
be added to the Subscribers list of the new message. In that case the
originating process will get either a succeed for the movement or a fail
for the movement.