Move player using only player and direction atom
In the previous commit I stated that if a message chain ended before an
exit connected the player's room to a room in the specified direction
then we could return a fail when we reached the last process in the
chain. Right now I don't feel like adding player movement fail
conditions to erlmud_object (to specific for so general a module) so the
player itself will recognize the "success" of a {move, Player,
Direction} as a failure to translate to a real move command.

erlmud_player translates {move, Player, Direction} to
{move, Player, PlayersRoom, Direction}, then a matching exit will
further translate it to {move, Player, PlayersRoom, DestinationRoom}.

I had no idea I'd have multiple cases of message translation on the fly
so I'm curious to see how it turns out.

I removed the example section from the README because it was ugly.

I've changed erlmud_world to print out only the state of each named
gen_server rather than the entire status.

This is all getting very fragile and is untested.