Stop at 2nd exit; #procs record
Inserted maybe_attempt to check and see if the current object is an exit
that isn't attached to the first encountered room. Once an event is
generated the first room encountered will set itself as the first room;
after this all exits will check to see if they are connected to that
room before proceeding. If an exit is not connected to the first room
then we simply ... well, we will simply mark ourselves as done and move
on to the next process in the chain.

I created a procs record after adding the first room since four fields
in a raw tuple was getting unwieldy.