erlmud_object behaviour, fixes ...
Turned erlmud_object into a behaviour to make sure each of the object
modules (erlmud_item, erlmud_player, etc.) implemented all of the
necessary callbacks. This also turns all the handle() functions in the
concrete object modules into attempt(), succeed() and fail().

Moved add and remove to erlmud_object since it's just adding a {Type,
Pid} to the list of props. Added added() and removed() callbacks in case
they are needed. I ignored YAGNI and I'll probably end up removing them.

Fixed the check for whether an exit should handle a message: I was
checking to see if the room _wasn't_ attached to the room rather than
whether it was. If it's attached to the first room that saw the message
then the exit can handle the message and pass it down the chain. If the
exit isn't attached to the first message that handled it then it just
returns a success since it's too far away to "hear" the event. e.g. if a
player tries to go south from one room to another, then every room
around the first room will hear about it but no rooms farther away.

Ditched the call() function in erlmud_object in favour of having
explicit succeed() and fail() calls. In order to support succeed() and
fail() with the call(...) construct I'd have to have multiple functions
with different arities and it would be ugly.

Added runtime tools and a debug tracing function in erlmud_world.

Added the pid to the erlmud_world state function so it's easy to see
which Pid has which state.