WIP: login & password states in conn FSM
Getting started with the login state and the password state in the
player connection FSM that the websocket handler will pass all input
to.

I'm thinking of having the conn FSM be a normal MUD process in the
process graph that can respond to events. For instance, if something
happens that requires player input the conn FSM would forward a message
out to the player and wait for a response before passing on the event or
failing it and returning the failure to the subscribed processes. e.g.
another player asks this player if they'd like to trade.