WIP: "database" and process index
In order to find a player to attach to a connection (erlmud_conn) I'll
need some way of looking it up. Each object (erlmud_object) will
register itself in the index with it's PID and identifier. The index is
a gen_server started by the top level supervisor.

Starting on some simple parsing so that erlmud_conn can either pass a
message to the player from the webpage or return a parse error
(currently just "huh?") back to the user.

A websocket handler will start a connection process which will, on
successful authentication, start a player process, store its PID and
then lookup the room the player should be in and connect it to that
room.