Add player to room on successful login
If a user enters credentials that match a player (stubbed out for now)
then the connection will send out an attempt to 'enter_world' for the
player to it's room and put the connection in the 'live' state. All
connected objects have a chance to refuse and/or subscribe to the
'enter_world' event, as per usual. If the 'enter_world' event succeeds
then the player adds itself to it's registered room and the connection
sends out a success message to the socket.

Currently erlmud_conn is doing double-duty as both the connection FSM
_and_ as a pseudo-erlmud_object: it uses the new erlmud_object:attempt
function to add itself as a subscriber to the event. It doesn't act as a
connected process but it does want to know if the 'enter_world' attempt
succeeds. This is ugly because I'm catching '$gen_cast' messages
manually. I'm thinking I could have another process that acted as the
connection erlmud_object that would proxy any requests from erlmud_conn.
Alternatively I could pass erlmud_object a function or have
erlmud_object itself provide a proxy for non-erlmud_objects that want to
be notified.

Player erlmud_objects are now going to hang out in "stasis" with a live
process, holding the pid of their room, but not receiving any messages
form the room. This allows some flexibility for logged-off players to
still be active in the mud but without direct user input. This also
allows me to get away without a database for the time being since there
is no "dead" state to be persisted. Once the user is logged in then the
player is added back into their current room.

Fixed erlmud_index so it registers itself.