WIP: Adding attack capability
An attack is a process that goes through a series of broadcast rounds
to:
a) calculate a hit score (> 0) is a hit
b) calculate a damage (< 1) means no damage
c) how long to wait to attack again

During each round of broadcasts all affected parties will add or
subtract their modifiers from the hit and damage scores to reveal a
final total.

Need to add checks to make sure a subsequent attack fails if the target
process is non-existent or dead.