AI module, Logging, Attacking /w tests
Created an module for computer controlled entities called AI.

Added a logging module and converted all io:format and existing
log calls to use erlmud_event_log. The log functions automatically
add the module name and erlmud_event_log automatically adds the calling
process' PID and--if found--the name of the PID registered with
erlmud_index. erlmud_index allows looking up an ID or a PID.

erlmud looks for the log path in $ERLMUD_LOG_PATH

Implemented attacking:

Attacking happens in several stages:
1) find player
2) start attack
3) find target
4) calculate "hit score" (i.e. did the attack land)
5) calculate damage
6) apply damage
7) calculate wait time
8) listen for message saying target is killed

Like all actions in erlmud, each of these steps is a message sent out to
all connected processes to either pass, fail, resend or modify. I've
added the ability to modify a message so that each connected process can
modify a value (e.g. hit score, damage amount) in a message. If the
processes were to send a new message then everyone would see it again.
This way processes only need to see the message once BUT they cannot
react to changes in the value as they have no guarantee which changes
they'll see: i.e. the first process to add an amount to the "hit score"
will not see any subsequent changes.

Currently the attack will cease if something decides that damage cannot
be applied, so this shouldn't be used.