Add "wield" /w test, streamline tests
Added a new erlmud_body_part module (name subject to change) that a user
can "wear" or "wield" things onto and "sheath" or "remove" things off
of. The particular verbs (e.g. "wear", "sheath", etc.) aren't in place
yet so far and erlmud_body_part just responds to "add" and "remove". The
verbs can be dependent on the player's body, the item, etc. so
add/remove in the code is clearest.

I'm using "stale" properties from the item being added so that the body
part can check if the item _can_ be added. I thought about having the
body part check first and then the item, but the body part will need to
know things about the item and vice versa. I then thought about having
one or the other send a list of constraints that would need to be
fulfilled ... but decided not to for some reason. Eventually I settled
(and "settled" is the right word) for using stale item properties that
may have changed since the message was sent. For instance, if the item
was gaining weight and was just light enough to be worn but then got
too heavy after the message was sent then a body part would have
something too heavy added. This will need more thought.

I streamlined the tests. Different tests now have smaller, more focused
"worlds" in which to work. Things were getting out of hand. I also made
it easier to test which objects have which values. I don't have a test
yet for removing an item from a body part.

I've got "fist" as an item and this will need either need to be changed
to a body part (and have body parts count as weapons) or added to a
hand.