Remove item by name without body part
Added the ability to remove an item from any body part on an object's
body using just the object's PID and the name of the item.

e.g. remove Player "bracelet"

The objects will pass that around until "bracelet" identifies itself
with a PID and the body part that owns the bracelet moves it back onto
its owner, the object.

This will lead to race conditions because there are three properties
being change asynchronously: the item's owner, the object's properties
(adding item) and the body parts properties (removing item). The body
part will be find because nothing else can happen to the body part while
it is busy moving crap around. The item and the owner could get screwed
up though: if the item is moved before this cast completes then it will
get moved back. It might be rare but given enough chances it will
happen, _especially_ when custom code triggers the movement of an item
(e.g. an enchanted item moving somewhere.)