WIP: hitpoint process, attempt(Owner, ...)
Adding a process that will manage hitpoint changes and send out a death
message if necessary. I can't remember why I put it into it's own
process.

I've added Owner to the attempt call for each object type since a lot of
objects are owned by other objects and only care about messages
concerning their owner. It was getting annoying to have a bunch of case
statements in attempt calls that duplicated the check if some PID
involved in the message was the owner of the current object.