Add attempt(Owner, ... add hp, life procs
Finished fixing the tests that broke when I added Owner to the params in
the attempt call (to enable pattern matching on it since it's used quite
often).

Finished moving HP logic and "alive/dead" logic to the hitpoints and
life processes. I still can't remember what motivated me to move the
hitpoints to their own processs. Maybe ... OH! Duh, I think so I could
share the hitpoint logic between both players and AI. I can see the
player and AI objects merging eventually: the main difference is where
input comes from; in the case of a player it comes from a human
through an erlmud_connection and in the case of an AI it's eventually
going to come from a separate process that will have an AI as it's
Owner.