WIP: add counter-attack
I linked the behaviour process to the current character attack process
and then added a "died" callback so that erlmud_object could listen for
dying processes and alert the "type" module. This would allow the
behaviour to update it's current "Owner's attack" to track whether or
not to counter attack.

I don't like the process linking though so I'm switching to having a
'stop_attack' attempt that the behaviour subscribes to. As much as
possible I'd like to stick to the attempt -> fail|succeed model instead
of adding side-channels and back doors for different situations.

I'm checking this in before I gut the 'died' callback in case I decide
to go back to it.