WIP: fixing tests after counterattack changes
The counterattack_behaviour test is the only test that is still failing.
The counterattack is happening but the player's attack process hasn't
shut down as I initially expected (after about three seconds thought).

Fixed a bug where the target property of an attack isn't replaced when
the name of the target is resolved to a pid.

Removed all the died event stuff, removed the process exit trap in
erlmud_object.

Set the starting damage to 0 instead of always assuming that at least 1
damage would be done.