WIP: Fix door locked move test, break others
The test process can subscribe to an attempted event but as soon as a
process resends the event all the subscriptions are blown away. The test
process isn't part of the graph so it doesn't get re-added.

Hmm, that triggered an idea: if a room 1 exit away from the original
source of the event resets the process list and resends, then it could
travel one more exit away from there. That's wrong.

To fix the door-locked test I simply tried to move with the door locked,
made sure the player didn't move, then tried with the door unlocked and
made sure the player _did_ move.