WIP: HTML + CSS logging
The test suite now creates a mock erlmud_object to receive gen_server
calls to get props. All test attempts go through this object. When the
receiver of the attempt logs the attempt erlmud_event_log tries to get
the props from the object sending the attempt. The mocked out object
will simply return an empty list.

All log messages need to be binary so that the list of logged terms can
be flattened. This allows us to get all the pids out and look up their
names with erlmud_index.