More work on HTML+CSS log "querying"
I'm still working out how to get the HTML logging to finish when the
test ends abruptly. In logs/log_host.html I'll have the whole HTML file
and then an import link to load whatever's been written to the log HTML
file.

This is a mess as I work out how to get this working.