pre-write HTML wrapper, test script
I tried loading the log output using an HTML import link but ran into a
cross-domain security roadblock; instead I added a line to the test
script that wraps the HTML log output in a header and footer.

The test is still failing but I'm closer to having a workable logging
solution even if it is a bit ridiculous.