Fix counterattack test, improve logging
The counterattack test needed longer for the change in hitpoints
property of the hitpoints process of the player. I shouldn't need two
seconds for this so I'll have to look into what's going on. Maybe add a
trigger or something so the MUD can tell the test when a certain event
has happened, or have the MUD shut itself down gracefully after a
certain amount of time.

Defaulted the display property of the columns and properties to none;
the checkboxes are out of sync but I got tired of having to uncheck the
boxes. Also changed the font for the log output so I can see more of it.

Took the attempt log messages out of the individual object modules and
put them all in erlmud_object. This makes them uniform and takes one
log call instead of 1 per object type.

All pids are now converted to their names from erlmud_index but their
pid also gets output: e.g. {<0.250.0>, pid_name}. I had to write a
custom flatten function to take apart message tuples and convert them to
lists with manual <<"{">>, <<",">> and <<"}">>. This is so I can get at
the pids in the message tuples.