WIP: fumbling around with searching
This was an attempt to get searching to work with any object including
enemies that might attack in response. I think I'm going to ditch that
for now and stick with inanimate or immobilized objects. I'm also
thinking of making as simple a search as possible as this MUD isn't
meant to be a pick-pocket simulator.