Revert to room limit, search WIP
I've reverted to using a the first adjacent room as the limit for
messages for now. Eventually I'd like to have custom limits so that
different actions propagate less or more but for now I don't need it.
Originally I was going to use it to constrain searches to a parent
object using paths or something. Now I'm going to build a hierarchy of
parent-child relationships so that when a search message succeeds each
subscriber can see if it's in the hieararchy; if it is then it knows
that it was searched and can fire off a command to print out it's
description. This will be very intensive with messages flying everywhere
but probably not anywhere near intensive enough for a modern machine to
even blink.

I've started aiming at a simple, uninteruptable search of immobile
objects only. I don't want to go near searching characters that can move
around or counter-attack because the complexity is more than I need
right now. Instead I plan to have containers determine if they are
searchable. A character ... well, I guess a character will have to know
that it's bound so that it doesn't allow itself to be searched.

Searching will only find the first container within a container. For
example, if you search a room that search will not propagate down to all
characters and containers in the room. If a room contains a box and a
character then those two objects will be described. If the player wants
to search a container within a container I'll have to provide a way for
the player to specify either a unique identifier or a path.

The counter-attack test is still broken even though I thought I'd fixed
it at one point.