Add lists of processes in props when merging
When an object passes on a message it picks the next process from its
list of properties. This change will look in properties that contain
lists of processes. Processes can be grouped together under a single
property such as "behaviours" instead of needing a specific property.

This fixes the counterattack_behaviour test.