WIP: Simple searching
Switching to a simpler search where only the top level children in a
"context" (e.g. room, character) will describe themselves. The child
items will send their descriptions back to their parent context to be
modified by the parent.

If an object describing itself on its own (e.g. look at Bob) then
it can just send "Bob is a human <description of non-process properites>;
however, if the object is seen in the context of a parent then we need to
send its description back to the parent so the parent (e.g. a room) can
modify the description with something like:
"In <room name> you see <character description".

You have to look at something to get it's child property descriptions. That is,
when you look directly at something you'll get a description of all it's children
one level deep.

e.g. look at the room:
> look
> You are in a featureless box.
> In the featureless box you see a human male
> In the featureless box you see a sword.
> In the featureless box you see a loaf of bread.
> look human male
> The human male is blah blah blah
> The human male is carrying/wearing/holding a hammer <-- hammer in context
> The human male is carrying/wearing/holding a funny hat
> look funny hat ... etc. etc.

I may even ditch the idea of searching for now and just use "look".