WIP: looking - send desc. to parent for framing
If we want to describe something that is a child object of something
else then we'll need to describe that parent child relation in the
description: e.g. "Bob is carrying a steel sword with a wire-bound
handle". We'll send the description of the sword back to Bob and tell
bob what property the description is from. Bob should be able to find
the matching child in the properties list, figure out what kind of
property it is and writing the proper description fram.