WIP: testing look command
Since there's no change in properties in any object when a player looks
at something I'm testing using a test socket. The test socket will set
up a connection, send the look command and save any output.

I'm having to work out how a player object gets started up and attached
to a room.

The socket connection will start a erlmud_object with erlmud_conn_obj to
subscribe to send events to send to the erlmud_conn object to relay to
the socket.