WIP: more work on socket -> conn -> conn obj
Instead of having a connection FSM that tracks the state of the user's
connection _and_ participates in the object graph I'm separating the
two: erlmud_connection FSM tracks the state of the user's connection
and erlmud_conn_obj participates in the MUD object graph subscribing
to messages and passing on relevant messages to the connection FSM.

I need to work in the "enter_world" subscription/result so that the room
is responsible for determining if the player can be added to the room.
Since erlmud_conn_obj is in the graph it will trap if "enter_world"
fails and pass it on to erlmud_conn to drop the socket connection and
clean up.

I figure all the players can be processes when the MUD starts up since
proesses are very cheap. I'll still need a data store at some point
(even if only text files, for easy editing) but for now I can assume
that if the MUD is running then all player processes are alive. ...
Except I'll eventually need to create new player processes. But again,
if a player can login then we just asssume the process exists and we
don't need to worry about reading in all the properties of the player
yet.

I'm not sure how I'm going to cobble together the graph connections
between the room, the player and the conn_obj. The connection might fire
up the conn_obj and the conn_obj can fire off the message to enter
world. If the event succeeds then:
- room stores player
- player stores room
- player stores conn_obj
- conn_obj stores player
The conn_obj is automatically subscribed to any message that it kicks
off so if it gets back a success message then it must be for the right
player.
The player will somehow need to be connected to the graph in order to
get the "enter_world" success message. Hmm, the conn_obj might need
to look up the player first and connect it.