WIP: searching
Still working on splitting the raw connection from the object that will
represent the connection in the object graph. The connection object
can't store the raw connection as {conn, Pid} because erlmud_object will
see that as a process in the graph: I've changed it to store it as
{conn, {Pid}}. The raw connection sends a message into the graph by
sending the connection object process a raw message of {Pid, Message}
and erlmud_object picks that up and injects it as an attempt into the
graph.

It's still not working but at least the player is now moved into the
room. (I also moved the Player character out of the room in the initial
test setup.)

The socket isn't getting any messages so I need to keep working on it.