Look works, tests pass!
Finally, after a bunch of false starts and a lot of "hammock" time I've
got the look command propagating down from a character to it's body
parts and finally any items on those body parts.

I had to clean up the test world (World 7 currently) to put the items on
the giant (instead of the player) as well as add names to the items.

erlmud.app.src has "look templates" that are a list of raw binary text
and property tuples. Each property tuple gets replaced with the value of
the corresponding property in the object being described.
(I need to remove the context verbs as I'm not using those now;
characters don't hold objects in their hand, they simply have an
floating inventory for now. I need to add inventory slots.)

How looking at a character works:
- The character outputs their description as per the "look template"
  in erlmud.app.src
- The character's body parts see that the character has been looked at,
  describe themselves (with .app.src template) and then _resend_ that
  look message with themselves as the target. This is _not_ so that
  the body_part can describe itself (which it's already done) but rather
  so that the items owned by the body_part can see the message to their
  owner and describe themselves.
- Items belonging to body parts see a look message to their body_part
  parent and describe themselves according to their template.

Currently the body part and item templates are very simple.

There's several different scenarios where a character can describe
themselves: looking at a room and seeing the character, but not its body
parts or items; looking at a character and seeing its body parts and
items; looking at an item in a room. Eventually I may have a way of
looking at an item on a character.

I'm going to need a way of giving objects unique identifiers within
their context. This might be easiest if the names are universally unique
and perhaps based on there object type to distinquish them. For example
a sword item might be i54: "i" for item and 54 as a unique item
identifier.