Update the README.
I don't feel like the README was adequately explaining what was
interesting about erlmud. After reading Craig Everett's (zxq9)
write-up about his erlmud I decided it was time to clarify my erlmud.

I'm guessing that there are technical names for different aspects of how
I'm building erlmud but as usual I'm enjoying stumbling around in the
dark re-inventing the wheel.

This change includes a tl;dr that explains erlmud in point form so
people can get the gist of it right away. It contains an introduction to
further explain the points, some examples to clarify and then a "why"
and "how" which delve into the details.

Explaining erlmud was overdue so I could remind myself why I chose to
build it the way I did and what the alternatives were.