Add character details; desc character in room
Added details to the test character so that when we describe it the
gender, height, weight etc. don't show up as ??.

If a player looks at room all characters in the room should trigger on
that event and describe themselves. Previously the character was only
describing itself if the it wasn't the source of the look command.
Actually, I may need to re-think this. A player looking at a room
shouldn't get back a description of their own character.