Finished look_room test
Finished the look_room test so that characters are described, but not
their body parts or items. Items are also described. I added an item to
World 7 that wasn't owned by the giant to be sure room items were being
picked up.

I switched characters from having a "room" property to having an "owner"
property because it's easier if all objects have an owner. If some
objects have specific types as owners it makes the code more complicated
and it implies that those objects can _only_ have those types of owners.
For example if characters have rooms (as owners) instead of just an
owner property then it implies that characters must be owned by rooms.

I also spruced up the items in World 7 with some descriptions to go with
their names instead of having ?? show up in the descriptions.