Create attributes that describe objects
Start of work to create attributes such as height, weight, gender, race
and maybe even strength, charisma, etc. These contribute to the
top-level description of an object. I think I'm going to ditch the
description templates and have all the attributes glom on their
descriptions to the {describe, Source, Target, ...} message. Once the
message succeeds the Target object can print out all the attribute
descriptions.

I add a "HandledBy" variable to add the PIDs of objects that had already
added their descriptions but this implies continuously re-sending the
event with new information added and that sounds odd. I should just be
able to have each attribute send the message as it goes by.

One problem with having attributes add their descriptions to the
describe event for their owner is that I can't control those attribute
descriptions. A character with a better perception skill, or some spell,
or other modifier won't be able to control how an attribute is
described.

For example, a mechanic with an advanced skill in calculating weights
might see that a large robot is roughly 2,439 kg, where a librarian with
little skill in calculating weights might guess at roughly 4,000 kg. It
might be better to have each attribute spit out it's description as it's
own event only when the object is directly observed:

> look
> You are standing in a ... blah blah room description.
> Room -> giant
> Room -> robot
> Room -> apple
> look giant
> Giant
> Giant -> 4m tall
> Giant -> 400 kg
> Giant -> male
> Giant -> wounded arm vs. Giant -> deep gash on arm
> etc.

In fact, I like this style as it makes me think of an old computer
printing out single facts on a dot matrix printer or something.