Attributes contribute to owner desc
Got the attributes to contribute to the description of the owner by
adding their descriptions to the {describe, Source, Owner, Context,
Descs} message as it goes by.

The attributes have a description property ('desc') that has a list of
binaries with a 'value' atom as a marker for where to insert the value
of the attribute.

Currently the attributes leave a trailing ", " but I'm not going to
worry about that for now.

Switched from a 'look' message to a 'describe' message since 'look' is
the initial action but then 'describe' is the resulting command. It
helps to have smaller sets of patterns for different messages. Too many
'look' message attempt and succeed patterns are confusing. Both
body_parts and attributes use the same {describe, Source, Owner,
Context, SubDescs} message but the body parts ignore the SubDescs; the
SubDescs are there for objects that contribute to the parent
description. Body parts do not contribute; they print out their own
message. I suppose I could have the body parts simply list themselves in
the owner description and then have items on body parts print themselves
out with the body part context.