WIP: backing out attribute descs added on attempt
I was adding attribute descriptions to the parent object during the
attempt phase. This would prevent attributes from being blocked or
manipulated. For example, if a spell made the character look taller
it would need to receive the attribute description event to change it.
If darkness prevented an attribute from being seen then the darkness
would need to block it.