Break logic into discrete modules
Pattern matching is getting very hard to understand with the amount of
logic that I'm putting into the various object type modules. Instead I'm
putting the logic into discrete handler modules. When an object receives
an "attempt" event each of its logic handlers will handle the attempt in
turn. The chain will short circuit on the first handler that doesn't
return a "not interested" flag (currently 'not_interested').

I'm leaning towards not having "object type" modules at all (e.g. room,
character, exit, etc.) but rather having the logic as part of the
properties of the object. A room would behave as a room not because it
was inately a room, but because it's attempt, succeed and fail handlers
include room logic handlers.

I'm going to need to have custom logic eventually anyway. Unless I want
to have an unending parade of unique erlmud_<foo> object type modules I
need a way to package the logic that makes characters different.