erlmud_attempt, _success and _failure
Changed erlmud_handler[...] to erlmud_attempt so that
erlmud_success[...] and erlmud_failure[...] would make more sense.

The type modules continue to look useless. However, having the logic
only tied together in configuration is going to be confusing. MUD
elements that will need to be configured correctly to have attempts that
line up with successes and failures. I may need to find some way to
check that logic is wired up correctly, or have a module that ties
things together. So maybe a type module that expresses by name a
combination of attempt, success and failure logic.
"erlmud_attack_player", "erlmud_attack_pitbull", "erlmud_open_door".
Maybe a module that takes a list of object propperties and adds the
appropriate attempts, successes and failures and can be chained together
with other "property logic adders" or something.

Maybe a better way would be to group together the attempt, success and
failure logic in a single module. This might mean some redundancy, but
that could be broken out into discrete modules that are packaged.

Instead of calling out to all the attempt handlers for an object on
attempt, and then success handlers on success, etc. The object could
just call out to its one list of handlers for any event regardless of
whether it's an attempt, success or failure. Each handler would have
three different functions.