More conversions to all-in-one handlers
To recap, I'm going from generic object modules to separate modules for
attempts, successes and failures, and finally converting
objects to collections of "all-in-one" handlers that cover attempts,
successes and failures in one module covering a specific event in a
specific way.

Object modules:
These were modules that reprenented an entire "type" (or class, if
you're into OO) of objects: rooms, characters, items, etc. The problem
with this is that there are no instances of the type: there's no way to
give a specific instance a particular behaviour without creating a whole
new type. i.e. there's no inheritance or extension. Also, it became
really confusing when all the event handling code was mashed together in
function clauses spanning the three stages of event handling: attempt,
success and failure. I first thought of wrapping each function body in a
function with an expressive name, but these quickly went from
"behaviour" or "event handler" functions to modules.

Event Stage-specific Handlers:
So I started creating custom behaviours that you could link to a
general object. However, it turned out that once you take out all the
behaviours out of an object type you're left with nothing but properties.
We already have all the properties in a proplist so we don't need an object
for that. Also, I was separating behaviours into separate modules for
handling the attempt, success and failure of an event. This made it _really_
confusing to follow the logic of how an attempt was handled. It might
provide less redundant code because small pieces could be
interchangeable, but the maintence cost and confusion would be too great
by far.

All-in-One Event Handlers:
So now I'm creating "handlers" which are modules that codify how to
respond to the attempt, success and failure of an event. This makes it
easy to follow the logic of a particular event and makes it easy to
create custom handlers for a particular event. Now when you look at a
particular event, such as a move, you can easily tell that all the event
code in the module involves moving. This is important for things like
descriptions which have a lot of back-and-forth events travelling
between different objects and looping over descriptive sub-objects.