Convert erlmud_conn_obj to event handlers
Create move, enter_world and send event handlers to replace the erlmud
connection "object". There is an erlmud_connection that communicates
with the client and then there _was_ an erlmud connection "object" that
lived in the object graph and participated in event handling.

There will still be an erlmud connection "object" but it will now be a
list of properties. This list is currently just going to be the three
event handlers: send, move and enter world.