Remove old attempt/success handlers, update world
The old erlmud_attempt_* and erlmud_success_* modules were from when I
was goign to have the different event handling stages (attempt, success,
failure) in different modules. This was way too confusing so I've
switched to having "all-in-one" handlers.

Updated the test world to have a bunch of hanlder macros for the
different types of objects. These will be added to the property lists.
erlmud_object looks for the "handlers" property to figure out what
handlers to route the event through.