Fixing build errors
Build the handler behaviour module first so that it exists when other
modules get built that rely on it.

Fix a bunch of old 'attempt' module names that are in all-in-one handler
files.

Remove/comment out some of the old add & remove calls that relied on
synchronous calls to objects to add or remove things. Instead when
something is added, removed, gotten, or dropped, both the old owner and
the new owner will be part of the event. The old and new owners will
both watch for a success and update their properties.