More handlers work ...
Fixing missing handlers in test world.
All objects need to have logic to handle events. All the objects in the
test world are there for a reason and will, probably, need logic. Some
might just be there to make sure something _doesn't_ happen (e.g. make
sure the shield isn't dropped when "drop sword" is entered) so they
should still have logic. Objects can now have have custom handlers for
only one type of event: e.g. custom attack logic but regular inventory,
move and look logic.

Change return from uninterested handlers to 'undefined'.
The erlmud_object module will pass attempts to handlers until the first
handler returns a value other than 'undefined'. Only one handler from an
object should be handling an attempt: e.g. if an attempt is for a player
to drop a sword then only the erlmud_handler_char_inv handler on the
character should be handling the event. Having two handlers having the
event will conflict. If all handlers return 'undefined' then
erlmud_object will default the return to {succeed, false, Props}; in
other words, "no change, no subscription".

Fix character, room and item inventory code.
I'm moving to some more expressive inventory management code to move
items to and from characters and rooms. The new messages are {move,
item, from, Source, to, Target}. The 'item' atom is so that rooms when
they are storing an item vs. a character, e.g. {Type, Pid}. This way
it's much easier to read what's going on than {Action, Pid1, Pid2,
Pid3}. Getting the order right required way too much thought. Eventually
I may need to formalize a protocol for object messages that I can test
for.

Change "is_attached_to_room" code. Instead of having a synchronous call
to an exit to see if it's attached to a room I'm going to make this
asynchronous and handle it ... I think from the exit itself. Anyhow, I
need to clear out the synchronous calls. Everything needs to stay
concurrent and asynchronous. This is a WIP and probably breaks things
the way it is.

Add in missing '_Owner' elements in attempt message tuples.
When all the logic for objects was in an object module I had some
shortcut functions that ignored the Owner element in the attempt message
tuple. I've removed that and put the _Owner "ignore this value"
placeholder back in. I might put in back in later but as I was breaking
apart this logic it was easier to keep all the handler attempt functions
the same.

Added erlmud_object:set/2 back in.
This is used in testing and is handy. It's synchronous so it should only
be used in testing.

Add note about not changing properties on attempts.
No handler attempt functions currently change the properties of their
object. They only change the values in the attempt message. However, at
some point there may be a property that needs to change in response to
an attempt. YAGNI _would_ apply except I've already coded it and I don't
want to introduce bugs taking it out.
(YAGNI: you aren't going to need it. i.e. Don't code for things you
_might_ need in the future because most of the time ... you aren't going
to need it. Just stick to what you _know_ you need and get _that_
working.)

Commented out tests.
Whoops. This was just for working on a particular test.